Disable button conditionals

+++ main.js
@@ -26,8 +26,15 @@ clearAll.addEventListener('click', clearEverything);
 listArea.addEventListener('click', deleteSidebarListItem);
 
 function onLoad() {
-
+  // buttonDefaultDisable();
 }
+
+// function buttonDefaultDisable() {
+//   addTaskItemButton.disabled = true;
+//   makeTaskListButton.disabled = true;
+//   searchButton.disabled = true;
+//   filterUrgency.disabled = true;
+// }
 
 function addListItemToSidebar(newListItem) {
   var listText = `
@@ -73,27 +80,31 @@ function clearInputFields() {
 }
 
 function addItemsToArray() {
-  var newListItem = new Items(taskItemInput.value);
-  taskItemsArray.push(newListItem);
-  console.log(taskItemsArray);
-  addListItemToSidebar(newListItem);
+  if (taskItemInput.value) {
+    var newListItem = new Items(taskItemInput.value);
+    taskItemsArray.push(newListItem);
+    console.log(taskItemsArray);
+    addListItemToSidebar(newListItem);
+  }
 }
 
 function onSubmitBtnClick() {
   createTodo(taskItemsArray);
 }
 
-function createTodo() {
-  console.log(taskItemsArray);
-  var newTodoCard = new Card(taskTitleInput.value, taskItemsArray);
-  console.log(newTodoCard.taskList.length);
-  todoCardsArray.push(newTodoCard);
-  console.log(todoCardsArray);
-  newTodoCard.saveToStorage();
-  appendCardToDOM(newTodoCard);
-  deleteAllListItemInSidebar();
-  clearInputFields();
-  taskItemsArray = [];
+function createTodo(taskItemsArray) {
+  if (taskTitleInput.value && listArea.innerHTML === !'') {
+    console.log(taskItemsArray);
+    var newTodoCard = new Card(taskTitleInput.value, taskItemsArray);
+    console.log(newTodoCard.taskList.length);
+    todoCardsArray.push(newTodoCard);
+    console.log(todoCardsArray);
+    newTodoCard.saveToStorage();
+    appendCardToDOM(newTodoCard);
+    deleteAllListItemInSidebar();
+    clearInputFields();
+    taskItemsArray = [];
+  }
 }
 
 
@@ -131,4 +142,43 @@ function iterateThruTasks(newTodoCard) {
       </li>
       `
   } return taskListIteration;
+}
+
+// ------------------------------------------------------------------------------------------
+
+function restoreMethods() {
+  var oldCollection = ideaCollection;
+  var newInstances = oldCollection.map(function(datum) {
+    datum = new Idea (datum.id, datum.title, datum.body, datum.star, datum.quality);
+    return datum;
+  }); 
+  ideaCollection = newInstances;
+  restoreCards(ideaCollection);
+}
+
+function restoreCards(ideaCollection) {
+  ideaCollection.forEach(function(datum) {
+    displayIdeas(datum);
+  });
+}
+
+function displayIdeas(ideaInstance) {
+  var ideaCard = `
+    <div class="card" data-id="${ideaInstance.id}">
+        <section class="cards__top card--section">
+          <div class="cards__top--left" alt="star-rating"></div>
+          <div class="cards__top--right" alt="delete-X"></div>
+        </section>
+        <section class="cards__middle card--section">
+          <h3 class="cards__middle--title" id="editable-title" contenteditable="true">${ideaInstance.title}</h3>
+          <p class="cards__middle--text" id="editable-body" contenteditable="true">${ideaInstance.body}</p>
+        </section>
+        <section class="cards__bottom card--section">
+          <img class="cards__bottom--left" src="images/upvote.svg">
+          <p class="cards__bottom--text">Quality: ${ideaInstance.quality}</p>
+          <img class="cards__bottom--right" src="images/downvote.svg">
+        </section>
+      </div>`;
+  cardTable.insertAdjacentHTML('afterbegin', ideaCard)
+  hidePrompt();
 }
